requirements.txt
fastapi
uvicorn
python-multipart
pydantic
python-dotenv
httpx
backboard-sdk
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-logging